:sparkles: add permissions to users api

+++ users/api.py
@@ -5,10 +5,18 @@ from rest_framework.viewsets import GenericViewSet
 from rest_framework.response import Response
 from rest_framework.generics import get_object_or_404
 
-from users.serializers import UserRegisterSerializer, UserSerializer
+from users.permissions import UserPermission
+from users.serializers import UserRegisterSerializer, UserSerializer, UserListSerializer
 
 
 class UserViewSet(APIView):
+
+    permission_classes = [UserPermission]
+
+    def get(self, request):
+        users=User.objects.all()
+        serializer = UserListSerializer(users, many=True)
+        return Response(serializer.data)
 
     def post(self, request):
         serializer = UserRegisterSerializer(data=request.data)
@@ -22,13 +30,17 @@ class UserViewSet(APIView):
 
 class UserDetailViewSet(APIView):
 
+    permission_classes = [UserPermission]
+
     def get(self, request, pk):
-        user = get_object_or_404(User, username=pk)
+        user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user)
         return Response(serializer.data)
 
     def put(self, request, pk):
-        user = get_object_or_404(User, username=pk)
+        user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         serializer = UserSerializer(user, data=request.data)
         if serializer.is_valid():
             updated_user = serializer.save()
@@ -38,7 +50,8 @@ class UserDetailViewSet(APIView):
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, pk):
-        user = get_object_or_404(User, username=pk)
+        user = get_object_or_404(User, pk=pk)
+        self.check_object_permissions(request, user)
         user.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
